feat: move info and status endpoints to admin prefix
