+++ server/misc/routers/root.js
@@ -14,18 +14,6 @@ const router = express.Router()
 
 const remoteServices = config.remoteServices.map(s => ({ ...s }))
 
-let info = { version: process.env.NODE_ENV }
-try { info = require('../../BUILD.json') } catch (err) {}
-router.get('/info', (req, res) => {
-  if (!req.user) return res.status(401).type('text/plain').send()
-  res.json(info)
-})
-
-router.get('/status', (req, res, next) => {
-  if (!req.user) return res.status(401).type('text/plain').send()
-  if (!req.user.adminMode) return res.status(403).type('text/plain').send(req.__('errors.missingPermission'))
-  status.status(req, res, next)
-})
 router.get('/ping', status.ping)
 
 router.get('/api-docs.json', (req, res) => {
